requirement.txt
patchright
pandas
ipynb
ipykernel
PyYAML